Reorganize into small and large stars and clean up imports
