+++ src/components/Star/Large/index.ts
@@ -1,0 +1,1 @@
+export { LargeStar } from './LargeStar'
