+++ src/components/Star/Small/SmallStar.tsx
@@ -1,0 +1,6 @@
+import React from 'react'
+import styled from 'styled-components'
+
+import { SmallStarAnims } from './anims'
+
+export const SmallStar = () => {}
