+++ src/components/Star/Small/index.ts
@@ -1,0 +1,1 @@
+export { SmallStar } from './SmallStar'
