--- src/components/Star/styles/index.ts
@@ -1,3 +1,0 @@
-import { Star } from './Star'
-
-export const Styled = { Star }
